% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dienfar.R
\name{fars_map_state}
\alias{fars_map_state}
\title{mapping datapoints and state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing each US state}

\item{year}{A string of length 4 which gives the year}
}
\value{
returns a map of the given state and the datapoints
}
\description{
mapping datapoints and state
}
