% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dienfar.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize the data selected by the fars_read_years function}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A string of length 4 which gives the year(s),in the fars filename with month and year attributes.}
}
\value{
returns a list of the summarized data for one or more years
}
\description{
Summarize the data selected by the fars_read_years function
}
