% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dienfar.R
\name{make_filename}
\alias{make_filename}
\title{Create custom filename with certain year#'}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string of lenght 4 which gives the year added to the filename.}
}
\value{
This function returns a string with the filename customized with given year
}
\description{
Create custom filename with certain year#'
}
\examples{
make_filename(2015)
}
